requirements.txt
openai>=1.40.0
streamlit>=1.37.0
python-dotenv>=1.0.1
plotly
